pyproject.toml
[project]
name = "shiva"
version = "0.4.16"
description = "A modern, async Python framework for distributed systems and microservices."
readme = "README.md"
requires-python = ">=3.12"
authors = [
  { name = "Boris Soldatov", email = "djsaff@gmail.com" }
]
license = "MIT"
dependencies = [
    "aiomisc>=14.0.0,<15.0.0",
    "aio-pika>=9.0.5,<10.0.0",
    "aioredis>=1.3.1,<2.0.0",
    "asyncpg>=0.28.0,<1.0.0",
    "PyYAML>=6.0.1,<7.0.0",
    "orjson>=3.5.2,<4.0.0",
    "fastapi",
    "argparse>=1.4.0,<2.0.0",
    "uvicorn>=0.13.4,<1.0.0",
    "loguru>=0.7.2,<1.0.0",
    "sentry-sdk>=1.0.0,<2.0.0",
    "flatdict>=4.0.1,<5.0.0",
    "typer>=0.3.2,<1.0.0",
    "art>=5.2,<6.0.0",
    "marshmallow>=3.13.0,<4.0.0",
    "motor>=3.6.0,<4.0.0",
    "aiobotocore>=2.16.0,<3.0.0",
    "packaging>=23.0,<24.0",
    "twine>=3.8.0,<4.0.0",
    "keyring>=23.5.0,<24.0.0",
    "uvloop>=0.17.0,<1.0.0; sys_platform == 'darwin' or sys_platform == 'linux'",
    "fastapi-restful",
    "prometheus-client>=0.18.0,<1.0.0",
    "mako>=1.3.0,<2.0.0",
    "starlette>=0.37.2,<1.0.0",
    "fastapi-utils[all]>=0.8.0,<1.0.0",
    "pytest>=8.4.1",
    "setuptools>=80.9.0",
    "pytest-asyncio>=1.0.0",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

[project.optional-dependencies]
dev = [
    "ipython>=8.22.1,<9.0.0"
]

[project.scripts]
shiva = "shiva.shiva_cli:main"

[project.urls]
homepage = "https://github.com/Zl0n1k/shivad"
bug-tracker = "https://github.com/Zl0n1k/shivad/issues"
documentation = "https://github.com/Zl0n1k/shivad#readme"

[tool.uv]
# You can add uv-specific configuration here if needed

[tool.ruff]
extend-include = ["*.ipynb"]

[tool.setuptools]
license-files = ["LICENSE"]

[tool.setuptools.packages.find]
where = ["shiva"] 
